visual flag
