+++ datahub-graphql-core/src/main/java/com/linkedin/datahub/graphql/resolvers/config/AppConfigResolver.java
@@ -4,24 +4,7 @@ import com.datahub.authentication.AuthenticationConfiguration;
 import com.datahub.authorization.AuthorizationConfiguration;
 import com.linkedin.datahub.graphql.QueryContext;
 import com.linkedin.datahub.graphql.featureflags.FeatureFlags;
-import com.linkedin.datahub.graphql.generated.AnalyticsConfig;
-import com.linkedin.datahub.graphql.generated.AppConfig;
-import com.linkedin.datahub.graphql.generated.AuthConfig;
-import com.linkedin.datahub.graphql.generated.EntityProfileConfig;
-import com.linkedin.datahub.graphql.generated.EntityProfilesConfig;
-import com.linkedin.datahub.graphql.generated.EntityType;
-import com.linkedin.datahub.graphql.generated.FeatureFlagsConfig;
-import com.linkedin.datahub.graphql.generated.IdentityManagementConfig;
-import com.linkedin.datahub.graphql.generated.LineageConfig;
-import com.linkedin.datahub.graphql.generated.ManagedIngestionConfig;
-import com.linkedin.datahub.graphql.generated.PoliciesConfig;
-import com.linkedin.datahub.graphql.generated.Privilege;
-import com.linkedin.datahub.graphql.generated.QueriesTabConfig;
-import com.linkedin.datahub.graphql.generated.ResourcePrivileges;
-import com.linkedin.datahub.graphql.generated.TelemetryConfig;
-import com.linkedin.datahub.graphql.generated.TestsConfig;
-import com.linkedin.datahub.graphql.generated.ViewsConfig;
-import com.linkedin.datahub.graphql.generated.VisualConfig;
+import com.linkedin.datahub.graphql.generated.*;
 import com.linkedin.metadata.config.DataHubConfiguration;
 import com.linkedin.metadata.config.IngestionConfiguration;
 import com.linkedin.metadata.config.TestsConfiguration;
@@ -143,6 +126,13 @@ public class AppConfigResolver implements DataFetcher<CompletableFuture<AppConfi
         entityProfilesConfig.setDomain(profileConfig);
       }
       visualConfig.setEntityProfiles(entityProfilesConfig);
+    }
+    if (_visualConfiguration != null && _visualConfiguration.getSearchResult() != null) {
+      SearchResultsVisualConfig searchResultsVisualConfig = new SearchResultsVisualConfig();
+      if (_visualConfiguration.getSearchResult().getEnableNameHighlight() != null) {
+        searchResultsVisualConfig.setEnableNameHighlight(_visualConfiguration.getSearchResult().getEnableNameHighlight());
+      }
+      visualConfig.setSearchResult(searchResultsVisualConfig);
     }
     appConfig.setVisualConfig(visualConfig);
