+++ datahub-graphql-core/src/main/resources/app.graphql
@@ -216,6 +216,11 @@ type VisualConfig {
   Configuration for the queries tab
   """
   entityProfiles: EntityProfilesConfig
+
+  """
+  Configuration for search results
+  """
+  searchResult: SearchResultsVisualConfig
 }
 
 """
@@ -248,6 +253,16 @@ type EntityProfileConfig {
   entity profile pages. If null, rely on default sorting from React code.
   """
   defaultTab: String
+}
+
+"""
+Configuration for a search result
+"""
+type SearchResultsVisualConfig {
+    """
+    Whether a search result should highlight the name/description if it was matched on those fields.
+    """
+    enableNameHighlight: Boolean
 }
 
 """
