+++ datahub-web-react/src/app/search/matches/SearchTextHighlighter.tsx
@@ -3,7 +3,8 @@ import Highlight from 'react-highlighter';
 import styled from 'styled-components';
 import { useMatchedFieldsByGroup } from '../context/SearchResultContext';
 import { useSearchQuery } from '../context/SearchContext';
-import { MatchedFieldName } from '../context/constants';
+import { MatchedFieldName } from './constants';
+import { useAppConfig } from '../../useAppConfig';
 
 type Props = {
     field: MatchedFieldName;
@@ -18,6 +19,8 @@ const StyledHighlight = styled(Highlight).attrs((props) => ({
 }))``;
 
 const SearchTextHighlighter = ({ field, text, enableFullHighlight = false }: Props) => {
+    const appConfig = useAppConfig();
+    const enableNameHighlight = appConfig.config.visualConfig.searchResult?.enableNameHighlight;
     const matchedFields = useMatchedFieldsByGroup(field);
     const hasMatchedField = !!matchedFields?.length;
     const normalizedSearchQuery = useSearchQuery()?.trim().toLowerCase();
@@ -27,7 +30,7 @@ const SearchTextHighlighter = ({ field, text, enableFullHighlight = false }: Pro
 
     return (
         <>
-            {hasMatchedField ? (
+            {enableNameHighlight && hasMatchedField ? (
                 <StyledHighlight search={hasSubstring ? normalizedSearchQuery : pattern}>{text}</StyledHighlight>
             ) : (
                 text
