+++ datahub-web-react/src/app/search/matches/matchedFieldPathsRenderer.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { MatchedField } from '../../types.generated';
-import { downgradeV2FieldPath } from '../entity/dataset/profile/schema/utils/utils';
+import { MatchedField } from '../../../types.generated';
+import { downgradeV2FieldPath } from '../../entity/dataset/profile/schema/utils/utils';
 
 export const matchedFieldPathsRenderer = (matchedField: MatchedField) => {
     return matchedField?.name === 'fieldPaths' ? <b>{downgradeV2FieldPath(matchedField.value)}</b> : null;
