+++ datahub-web-react/src/appConfigContext.tsx
@@ -27,6 +27,9 @@ export const DEFAULT_APP_CONFIG = {
         entityProfile: {
             domainDefaultTab: null,
         },
+        searchResult: {
+            enableNameHighlight: false,
+        },
     },
     authConfig: {
         tokenAuthEnabled: false,
