+++ datahub-web-react/src/graphql/app.graphql
@@ -45,6 +45,9 @@ query appConfig {
                     defaultTab
                 }
             }
+            searchResult {
+                enableNameHighlight
+            }
         }
         telemetryConfig {
             enableThirdPartyLogging
